refactor: Make Package for Member
