--- src/main/kotlin/com/skhu/mypack/auth/domain/enum/Provider.kt
@@ -1,5 +1,0 @@
-package com.skhu.mypack.auth.domain.enum
-
-enum class Provider {
-    NONE, GOOGLE
-}
