+++ src/main/kotlin/com/skhu/mypack/member/dao/MemberRepository.kt
@@ -1,6 +1,6 @@
-package com.skhu.mypack.auth.dao
+package com.skhu.mypack.member.dao
 
-import com.skhu.mypack.auth.domain.Member
+import com.skhu.mypack.member.domain.Member
 import org.springframework.data.domain.Slice
 import org.springframework.data.jpa.repository.JpaRepository
 
@@ -8,5 +8,5 @@ interface MemberRepository : JpaRepository<Member, Long> {
     fun findByEmail(email: String): Member?
     fun findByName(name: String): Member?
     fun findAllByNameLike(query: String): Slice<Member>
-    fun isExistsByName(name: String): Boolean
+    fun existsByName(name: String): Boolean
 }
