+++ src/main/kotlin/com/skhu/mypack/member/domain/enum/Provider.kt
@@ -1,0 +1,5 @@
+package com.skhu.mypack.member.domain.enum
+
+enum class Provider {
+    NONE, GOOGLE
+}
